feat: forward and back button for slider
