+++ web/src/components/UpcomingEvents.tsx
@@ -1,6 +1,12 @@
 import UpcomingEventsList from "./UpcomingEventsList";
 import { Event } from "../types/types";
-import { IoArrowBackCircleOutline, IoArrowForwardCircleOutline } from "react-icons/io5";
+import {
+  IoArrowBackCircleOutline,
+  IoArrowForwardCircleOutline,
+} from "react-icons/io5";
+import useScreenSize from "../hooks/useScreenSize";
+import { useState } from "react";
+import { useKeenSlider } from "keen-slider/react";
 
 interface UpcomingEventsProps {
   upcomingEvents: Event[];
@@ -8,33 +14,47 @@ interface UpcomingEventsProps {
 }
 
 const UpcomingEvents: React.FC<UpcomingEventsProps> = ({ upcomingEvents }) => {
+  const [currentSlide, setCurrentSlide] = useState(0);
+  const screenSize = useScreenSize();
+
+  const [sliderRef, instanceRef] = useKeenSlider({
+    initial: 0,
+    slideChanged(slider) {
+      setCurrentSlide(slider.track.details.rel);
+    },
+    mode: "free-snap",
+    slides: {
+      origin: "auto",
+      perView: screenSize.width > 640 ? (screenSize.width > 1000 ? 3 : 2) : 1,
+    },
+  });
+
   return (
     <div className="from-AUIS-dark-teal to-AUIS-teal bg-gradient-to-b pb-20">
       <h1 className="mx-3 py-10 text-center text-5xl font-bold text-white">
         Our Upcoming Events!
       </h1>
       <div className="mx-auto flex items-center justify-center">
-      <LeftArrow
-            onClick={(e: any) =>
-              console.log("clicked")
-              // e.stopPropagation() || instanceRef.current?.next()
-            }
-            disabled={false
-              // currentSlide >=
-              // instanceRef.current.track.details.slides.length - (screenSize.width > 640 ? (screenSize.width > 1000 ? 3 : 2) : 1)
-            }
-          />
-        <UpcomingEventsList upcomingEvents={upcomingEvents} />
+        <LeftArrow
+          onClick={(e: any) =>
+            e.stopPropagation() || instanceRef.current?.prev()
+          }
+          disabled={currentSlide === 0}
+        />
+        <UpcomingEventsList
+          sliderRef={sliderRef}
+          upcomingEvents={upcomingEvents}
+        />
         <RightArrow
-            onClick={(e: any) =>
-              console.log("clicked")
-              // e.stopPropagation() || instanceRef.current?.next()
-            }
-            disabled={false
-              // currentSlide >=
-              // instanceRef.current.track.details.slides.length - (screenSize.width > 640 ? (screenSize.width > 1000 ? 3 : 2) : 1)
-            }
-          />
+          onClick={(e: any) =>
+            e.stopPropagation() || instanceRef.current?.next()
+          }
+          disabled={
+            currentSlide >=
+            upcomingEvents.length -
+              (screenSize.width > 640 ? (screenSize.width > 1000 ? 3 : 2) : 1)
+          }
+        />
       </div>
     </div>
   );
@@ -45,7 +65,7 @@ function LeftArrow(props: { disabled: boolean; onClick: (e: any) => void }) {
   return (
     <IoArrowBackCircleOutline
       onClick={props.onClick}
-      className={`arrow mx-4 hidden h-16 w-16 text-gray-400 hover:cursor-pointer sm:flex ${disabled ? "opacity-65" : " "}`}
+      className={`arrow mx-4 hidden h-16 w-16 text-gray-400 hover:cursor-pointer sm:flex ${disabled ? "opacity-40" : "hover:scale-110"}`}
     />
   );
 }
@@ -55,7 +75,7 @@ function RightArrow(props: { disabled: boolean; onClick: (e: any) => void }) {
   return (
     <IoArrowForwardCircleOutline
       onClick={props.onClick}
-      className={`arrow mx-4 hidden h-16 w-16 text-gray-400 hover:cursor-pointer sm:flex ${disabled ? "opacity-65" : " "}`}
+      className={`arrow mx-4 hidden h-16 w-16 text-gray-400 hover:cursor-pointer sm:flex ${disabled ? "opacity-40" : "hover:scale-110"}`}
     />
   );
 }
